pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "rangeflow"
version = "0.1.0"
description = "Certified Robustness and Uncertainty Quantification for AI"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT License"}
authors = [
  {name = "Dheeren-Tejani", email = "dheerenntejani@gmail.com"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "numpy",
]

[project.optional-dependencies]
gpu = ["cupy-cuda12x"]
torch = ["torch>=2.0.0"]

[project.urls]
"Homepage" = "https://github.com/yourusername/rangeflow"